requirements.txt
requests
threading
ruamel.yaml
ftfy
pyOpenSSL
pandas
